% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_ts.R
\name{add_sat_family}
\alias{add_sat_family}
\title{Add the family to which the satelite belongs to}
\usage{
add_sat_family(x)
}
\arguments{
\item{x}{a tibble with time series of observations.}
}
\value{
the tibble \code{x} with a new column called \code{sat_family}.
}
\description{
Add the family to which the satelite belongs to
}
