% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brstate_tb}
\alias{brstate_tb}
\title{Metadata from Brazilian states}
\format{
\subsection{\code{brstate_tb}}{

A data frame with 27 rows and 5 columns:
\describe{
\item{code_state}{State ID}
\item{abbrev_state}{State name abbreviated to 2 letters.}
\item{name_state}{State name}
\item{code_region}{Code of the region to which the state belongs}
\item{name_region}{Name of the region}
}
}
}
\source{
\url{https://cran.r-project.org/package=geobr}
}
\usage{
brstate_tb
}
\description{
Metadata from Brazilian states which could provide additional levels of
aggregation.
}
\keyword{datasets}
