% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{component_analysis}
\alias{component_analysis}
\title{Component analysis of time series}
\usage{
component_analysis(train_tb)
}
\arguments{
\item{train_tb}{a tibble with time series data.}
}
\value{
a list of ggplot2 objects.
}
\description{
Produce figures that help diagnosing time series by their components.
}
