% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_bfast.R
\name{convert_to_ts}
\alias{convert_to_ts}
\title{Convert to time series}
\usage{
convert_to_ts(data_tb, n_freq = 12)
}
\arguments{
\item{data_tb}{a tibble::tibble object.}

\item{n_freq}{a interger(1). The number of element per cycle.}
}
\value{
a stats::ts object.
}
\description{
Utility function to convert the input tibble into a stats::ts object
compatible with bfast.
}
