% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{date_to_period}
\alias{date_to_period}
\title{Convert a date into a period}
\usage{
date_to_period(x)
}
\arguments{
\item{x}{a date object.}
}
\value{
a character.
}
\description{
Given a date (base::Date), return its period.
}
