% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download data from the Queimadas Program}
\usage{
download_data(
  out_dir,
  data_url = get_data_url(),
  wait_time = 5,
  overwrite_files = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{out_dir}{a character(1). Path to a directory for storing the
downloaded data.}

\item{data_url}{a character(1). The URL of the Web page where the data is
hosted.}

\item{wait_time}{a numeric(1). Number of seconds to wait between requests
for files.}

\item{overwrite_files}{a logical(1). Should the downloaded files overwrite
older versions?}

\item{quiet}{a logical(1). Should status messages be displayed during
download?}
}
\value{
a character. Full path to the downloaded files.
}
\description{
Download data files from the Queimadas Program.
}
