% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_ts.R
\name{filter_complete_series}
\alias{filter_complete_series}
\title{Filter complete series}
\usage{
filter_complete_series(x, n_obs)
}
\arguments{
\item{x}{a tibble with the variables period, satelite, and n.}

\item{n_obs}{number of observations in a period.}
}
\value{
a modified version of x.
}
\description{
Keep those series with the complete number observations for a period.
}
