% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_queimadas.R
\name{fit_lm}
\alias{fit_lm}
\title{fit a linear model}
\usage{
fit_lm(data_df, formula, clevel)
}
\arguments{
\item{data_df}{a data frame.}

\item{formula}{a character. A formula object (see \code{stats::formula}).}

\item{clevel}{a numeric(1). The confidence level for estimating the
confidence interval.}
}
\value{
a list with two objects: an \code{stats::lm} object and a the input data
frame with addicional columns.
the data used to fit the model.
}
\description{
Fit a linear model to the given data.
}
