% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_queimadas.R
\name{fit_lm_01_months}
\alias{fit_lm_01_months}
\title{Utilitary function to fit a linear model month by month}
\usage{
fit_lm_01_months(x_df, y_df, formula = "y ~ x", clevel = 0.95)
}
\arguments{
\item{x_df}{a data frame.}

\item{y_df}{a data frame.}

\item{formula}{a character. A formula object (see \code{stats::formula}).}

\item{clevel}{a numeric(1). The confidence level for estimating the
confidence interval.}
}
\value{
a list of 12 where each element is composed of a \code{stats::lm} object
and a data frame with the data used to fit the model.
}
\description{
Given two data frames, merge them and then fit a linear model for each
month, ignoring the observations' year.
}
