% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_queimadas.R
\name{fit_monthly_lm}
\alias{fit_monthly_lm}
\title{Utilitary function to fit a linear model by month}
\usage{
fit_monthly_lm(x_df, y_df, formula = "y ~ x", clevel = 0.95)
}
\arguments{
\item{x_df}{a data frame.}

\item{y_df}{a data frame.}

\item{formula}{a character. A formula object (see \code{stats::formula}).}

\item{clevel}{a numeric(1). The confidence level for estimating the
confidence interval.}
}
\value{
a list with two objects: an \code{stats::lm} object and a data frame with
the data used to fit the model.
}
\description{
Given two data frames, merge them and then fit a linear model using their
overlaping observations using the observations' month,  but ignoring
their year.
}
