% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_queimadas.R
\name{fit_montly_lm}
\alias{fit_montly_lm}
\title{Utilitary function to fit a linear model by month}
\usage{
fit_montly_lm(x_df, y_df)
}
\arguments{
\item{x_df}{a data frame.}

\item{y_df}{a data frame.}
}
\value{
a list with two objects: an \code{stats::lm} object and a data frame with
the data used to fit the model.
}
\description{
Given two data frames, merge them and fit a linear model using their
overlaping observations while and using the observations' month (ignoring
their year).
}
