% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_stats.R
\name{get_bottom_row}
\alias{get_bottom_row}
\title{Get an aggregated bottom row}
\usage{
get_bottom_row(data_tb, f, na_rm = FALSE)
}
\arguments{
\item{data_tb}{a data frame.}

\item{f}{a function.}

\item{na_rm}{a logical. A parameter passed to \code{f}.}
}
\value{
an one-row data frame.
}
\description{
Get a row with the results of applying a funcion to each numeric column of
the given data frame.
}
