% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{get_brazil_year_month}
\alias{get_brazil_year_month}
\title{Get fire data from Brazil by year and month}
\usage{
get_brazil_year_month(db_con, table_name)
}
\arguments{
\item{db_con}{a conection to a data base of queimadas.}

\item{table_name}{name of the table in the data base.

return a DBI object. This object requires a call to \code{dplyr::collect()} to
retrieve the actual data.}
}
\description{
Get data from the database coresponding to Brazil, excluding industrial
areas, aggregated by year and month.
}
