% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_ts.R
\name{get_break_lines_year}
\alias{get_break_lines_year}
\title{Get the yearly breaklines for plot}
\usage{
get_break_lines_year(data_tb, period_col = "period", break_pattern = "-08$")
}
\arguments{
\item{data_tb}{a tibble representing a time series.}

\item{period_col}{a character. Name of the column identifying the time
period.}

\item{break_pattern}{a character to identify the break points.}
}
\value{
a character.
}
\description{
Get the points to insert breaklines in for the figures (plots).
}
