% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{get_cdata}
\alias{get_cdata}
\title{Utilitary function to get a column from a data frame}
\usage{
get_cdata(x, cname)
}
\arguments{
\item{x}{a data frame.}

\item{cname}{a character. Name of an object (column or element) in \code{x}.}
}
\value{
a vector.
}
\description{
Extract the data from a a column in a data frame.
}
