% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{get_data_url}
\alias{get_data_url}
\title{Get the data URL}
\usage{
get_data_url()
}
\value{
a character(1). An URL.
}
\description{
Get the URL of the Web Page that holds the links to the data files.
}
