% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{get_latex_char}
\alias{get_latex_char}
\title{Convert a data frame into a LaTeX-sanitized character data frame}
\usage{
get_latex_char(x)
}
\arguments{
\item{x}{a data frame.}
}
\value{
a character data frame.
}
\description{
Convert a data frame to another in which the contents are converted into
character vectors compatible with LaTeX.
}
