% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{get_lm_equation}
\alias{get_lm_equation}
\title{Get the equation from an \code{lm} object}
\usage{
get_lm_equation(lm_obj, dig = 3)
}
\arguments{
\item{lm_obj}{an `lm`` object.}

\item{dig}{an integer. The number of decimal places for the model
coefficients.}
}
\value{
a character.
}
\description{
Get the equation of a linear model as a character.
}
