% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_lm.R
\name{get_lm_residuals}
\alias{get_lm_residuals}
\title{Get model's residuals}
\usage{
get_lm_residuals(lm_obj)
}
\arguments{
\item{lm_obj}{a linear model object (stats package).}
}
\value{
a numad numeric.
}
\description{
Get the model's residuals.
}
