% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_queimadas.R
\name{get_modis_collections_dates}
\alias{get_modis_collections_dates}
\title{Get the MODIS collections}
\usage{
get_modis_collections_dates()
}
\value{
a data frame (tibble) with the following columns:
\itemize{
\item id an integer.
\item satellite a character(1). Satellite (or platform) name.
\item sensor a character(1). Sensor's name.
\item collection a character(1). MODIS collection number.
\item from,to a Date(1) object. The dates of start and end of temporal coverage
of the collection.
}
}
\description{
This function returns a data frame (tibble) with data regarding MODIS
collections.
}
\seealso{
Summary of the differences between various versions of MODIS data
\verb{https://nsidc.org/data/modis/version-history}.
}
