% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{get_month_from_period}
\alias{get_month_from_period}
\title{Extract the month from the period}
\usage{
get_month_from_period(x)
}
\arguments{
\item{x}{a character representing an aggregation period (e.g. "2028-07").}
}
\value{
a chracter. A month.
}
\description{
Given a period (e.g. "2028-07"), get the month ("07").
}
