% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_sizes.R
\name{get_paper_size}
\alias{get_paper_size}
\title{Get paper size}
\usage{
get_paper_size(name, orientation = "ls")
}
\arguments{
\item{name}{name of the paper size (e.g. "A4").}

\item{orientation}{either landscape (ls) or portrait (pt).}
}
\value{
a names list.
}
\description{
Get the dimensions of a international standard paper size.
}
