% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_plot_forecast_vs_obs}
\alias{get_plot_forecast_vs_obs}
\title{Get a forecasting versus observations plot}
\usage{
get_plot_forecast_vs_obs(obs_tb, forecast_tb)
}
\arguments{
\item{obs_tb}{A data frame of observations.}

\item{forecast_tb}{a data frame of forecast values.}
}
\value{
A plot (ggplot2) object.
}
\description{
Get a ggplot2 plot which compares a forecast versus a set of observations.
}
