% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_plot_ref_sats}
\alias{get_plot_ref_sats}
\title{Get a regression plot between reference satellites}
\usage{
get_plot_ref_sats(x, y, data_df, lm_obj)
}
\arguments{
\item{x, y}{a character(1). Name of a reference satellite.}

\item{data_df}{a data frame with overlapping observations of reference
satellites.}

\item{lm_obj}{a regression object (see \code{stats::lm}).}
}
\value{
A plot (ggplot2) object.
}
\description{
Get a ggplot2 plot which compares a reference satellite observations.
}
