% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_plot_ref_sats_01}
\alias{get_plot_ref_sats_01}
\title{Get a regression plot by month between reference satellites}
\usage{
get_plot_ref_sats_01(x, y, data_df)
}
\arguments{
\item{x, y}{a character(1). Name of a reference satellite.}

\item{data_df}{a data frame with overlapping observations of reference
satellites.}
}
\value{
A plot (ggplot2) object.
}
\description{
Get a ggplot2 plot which compares a reference satellite observations in a
month-by-month basis.
}
