% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_plot_ts_with_bfast_break}
\alias{get_plot_ts_with_bfast_break}
\title{Get a plot showing BFast breaks}
\usage{
get_plot_ts_with_bfast_break(bfast_ts_tb, bfast_top_tb, break_lines)
}
\arguments{
\item{bfast_ts_tb}{A tibble with the time series, including a logical column
(\emph{after_break}) representing the observations before and after the break.}

\item{bfast_top_tb}{A tibble with only the top observations of each period
(year).}

\item{break_lines}{A Date. Dates on which the perio changes (years).}
}
\value{
A ggplot2 object.
}
\description{
Plot times series of queimadas data showing their break points.
}
