% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_queimadas.R
\name{get_queimadas_dates}
\alias{get_queimadas_dates}
\title{Get the dates of usage of the reference satellites}
\usage{
get_queimadas_dates()
}
\value{
a data frame (tibble) with data about the time intervals of the
reference satellites in the Queimadas program:
\itemize{
\item id an integer(1).
\item satellite a character(1). Satellite's name.
\item sensor a character(1). Name of the sensor on board of the satellite.
\item launch_date a Date(1) object.
\item from,to a Date(1). Date from which the reference satellite is (or was)
used as reference satellite by the Queimadas program.
}
}
\description{
This function returns the time interval in which the Queimadas program used
each of the reference satellites.
}
\seealso{
Queimadas program, frequently asked questions, questions 7, 38, and 39
\verb{https://terrabrasilis.dpi.inpe.br/queimadas/portal/pages/secao_informacoes/faq/index.html}
}
