% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_queimadas.R
\name{get_queimadas_web_table}
\alias{get_queimadas_web_table}
\title{Format a forecast to look like Queimadas web table}
\usage{
get_queimadas_web_table(data_tb)
}
\arguments{
\item{data_tb}{a data frame representing a Queimadas forecast.}
}
\value{
a data frame where each year correspond to a row and each column to
a month.
}
\description{
Given a forecast data frame, format it to look like the the tables used in
the Queimadas web page.
}
