% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_queimadas.R
\name{get_queimadas_web_table_latex}
\alias{get_queimadas_web_table_latex}
\title{Get a LaTeX table code from a Queimadas web styled table}
\usage{
get_queimadas_web_table_latex(data_tb, tab_label, caption)
}
\arguments{
\item{data_tb}{a data frame in the Queimadas style. Possibly resulting from
a call to \code{get_queimadas_web_table()}.}

\item{tab_label}{a character(1). Label for the LaTeX's reference.}

\item{caption}{a character(1).}
}
\value{
a character.
}
\description{
Take a data frame in the Queimadas web style and convert it the LaTeX code of
a simililar table.
}
