% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_queimadas.R
\name{get_sat_data}
\alias{get_sat_data}
\title{Utilitary function to filter observations by satellite name}
\usage{
get_sat_data(sat_name, data_df)
}
\arguments{
\item{sat_name}{a character. Name of a satellite.}

\item{data_df}{a data frame with satellite data.}
}
\value{
a data frame with the same column as the input data frame, but with
the same of fewer rows.
}
\description{
Given a data frame of satellite observations, return those that belong to a
specific satellite.
}
