% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_queimadas.R
\name{get_sat_data_cell}
\alias{get_sat_data_cell}
\title{Utilitary function to filter observations by satellite and cell}
\usage{
get_sat_data_cell(sat_name, path_row, data_df)
}
\arguments{
\item{sat_name}{a character(1). The name of a satellite.}

\item{path_row}{a character(1). The location of the the wanted observation.}

\item{data_df}{a data frame with observations.}
}
\value{
a data frame similar to the given one but with potentially fewer
observation.
}
\description{
Get an observation from the given data frame corresponding to an specific
satellite and path-row.
}
