% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{get_sat_pairs}
\alias{get_sat_pairs}
\title{Get pairs of satellites for analysis}
\usage{
get_sat_pairs(data_tb, satellites)
}
\arguments{
\item{data_tb}{a tibble with data from queimadas.}

\item{satellites}{a character with satellite names to include. Among their
names, it must be those of candidate satelltes.}
}
\value{
a tibble.
}
\description{
Get a data frame of pairs of satellites for analysis.
}
