% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{get_satellites}
\alias{get_satellites}
\title{Get satellites' names}
\usage{
get_satellites()
}
\value{
a character.
}
\description{
Get a character vector with the names of the reference satellites used by
the Queimadas program.
}
