% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_queimadas.R
\name{get_ts_overlap}
\alias{get_ts_overlap}
\title{Get overlap among time series}
\usage{
get_ts_overlap(x_df, y_df, cname)
}
\arguments{
\item{x_df, y_df}{a data frame representation of a time series.}

\item{cname}{a character(1). Name of a column in both input time series.}
}
\value{
a data frame.
}
\description{
Get the common time steps among time series
}
