% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{get_year_from_period}
\alias{get_year_from_period}
\title{Extract the year from the period}
\usage{
get_year_from_period(x)
}
\arguments{
\item{x}{a character representing an aggregation period (e.g. "2028-07").}
}
\value{
a character. A year.
}
\description{
Given a period (e.g. "2028-07"), get the year ("2028").
}
