% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grid_sf}
\alias{grid_sf}
\title{Grid used for spatial aggregation}
\format{
\subsection{\code{grid_sf}}{

A data frame with 988 rows and 6 columns:
\describe{
\item{id}{Identifier.}
\item{gid}{Geometry identifier}
\item{path_row}{Identifier build using the relative location of the cell in the grid.}
\item{orbita}{Path as an integer.}
\item{ponto}{Row as an integer.}
\item{geom}{Geometry.}
}
}
}
\usage{
grid_sf
}
\description{
This is the reference grid used for the spatial aggregation of data during
analysis.
}
\keyword{datasets}
