% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{is_period_valid}
\alias{is_period_valid}
\title{Is the period valid?}
\usage{
is_period_valid(x)
}
\arguments{
\item{x}{a character representing a period (e.g. '2007-01').}
}
\value{
a logical(1).
}
\description{
Check if the given period is valid.
}
