% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_df.R
\name{lnames2df}
\alias{lnames2df}
\title{Add the list's names to the data frames in the list}
\usage{
lnames2df(df_ls, cname)
}
\arguments{
\item{df_ls}{a named list of data frames.}

\item{cname}{a character(1). The name for the new column.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    A list of data frames.
}\if{html}{\out{</div>}}
}
\description{
Given a list of data frames, add each list's name to its corresponding data
frame.
}
