% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{load_database}
\alias{load_database}
\title{Load files to a database}
\usage{
load_database(
  db_con,
  file_paths,
  c_types = "ddTcccccdnncd",
  table_name = "fire_foci",
  disconnect = TRUE
)
}
\arguments{
\item{db_con}{a connection to a data base.}

\item{file_paths}{a character. Path to CSV files.}

\item{c_types}{a character(1). Data types of the columns in input files.}

\item{table_name}{a character(1). Name of the table to creater in the
database.}

\item{disconnect}{a logical(1). Should the database connection be closed?}
}
\value{
the connection to the database (invisible).
}
\description{
Load the given files to a database of fire foci.
}
