% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_queimadas.R
\name{overlap_len}
\alias{overlap_len}
\title{Get the overlap length among time series}
\usage{
overlap_len(x_df, y_df, cname)
}
\arguments{
\item{x_df, y_df}{a data frame.}

\item{cname}{a character. The name of a common column in both input data
frames, which representes the time steps.}
}
\value{
an integer(1). The number of common steps in the time series.
}
\description{
Given a pair of time series represented as data frames, return the number
of time steps of overlap among them.
}
