% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{period_to_date}
\alias{period_to_date}
\title{Convert from period to date}
\usage{
period_to_date(x)
}
\arguments{
\item{x}{a character.}
}
\value{
a date object (\code{base::Date}).
}
\description{
Given a period (e.g. '2007-01'), convert it into a date object.
}
