% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_lm.R
\name{predict_ci}
\alias{predict_ci}
\title{Predict including a confidence level}
\usage{
predict_ci(lm_obj, new_data = NA, clevel = 0.95)
}
\arguments{
\item{lm_obj}{a model (\code{stats::lm}).}

\item{new_data}{NA or a numeric vector. If a NA is given, then it will use
the data in the model for the prediction. Otherwise it will predict for each
element in the given vector.}

\item{clevel}{a numeric(1). The confidence level to use.}
}
\value{
a data frame.
}
\description{
Use the given model to predict including the lower and upper values at the
given confidence level.
}
