% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_queimadas.R
\name{predict_ci_01}
\alias{predict_ci_01}
\title{Predict month by month including a confidence level}
\usage{
predict_ci_01(lm_ls, new_data, clevel)
}
\arguments{
\item{lm_ls}{a list of models (\code{stats::lm}).}

\item{new_data}{a data frame with data to predict using the given models.}

\item{clevel}{a numeric(1). The confidence level to use.}
}
\value{
a list of data frames.
}
\description{
Use the given models to predict, including the lower and upper values at the
given confidence level.
}
