% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_bfast.R
\name{run_bfast}
\alias{run_bfast}
\title{Run BFAST}
\usage{
run_bfast(x)
}
\arguments{
\item{x}{a stats::ts object (see \code{convert_to_ts}).}
}
\value{
an bfast object.
}
\description{
Utility function to run BFAST.
}
