% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{test_normal_residuals}
\alias{test_normal_residuals}
\title{Test if the model residuals are normal}
\usage{
test_normal_residuals(train_tb)
}
\arguments{
\item{train_tb}{a tibble with time series data.}
}
\value{
a list of ggplot2 objects.
}
\description{
Get the distribution of model residuals to check if the follow a normal
distribution.
}
