% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{unzip_files}
\alias{unzip_files}
\title{Unzip files}
\usage{
unzip_files(files, out_dir)
}
\arguments{
\item{files}{a character. Path to zip files.}

\item{out_dir}{a character(1). Path to a directory.}
}
\value{
a character. Path to the unzipped files. The vector names are the
input zip files.
}
\description{
Unzip the given zip files into the given directory.
}
