% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{visual_test_ts}
\alias{visual_test_ts}
\title{Visual test of time series}
\usage{
visual_test_ts(train_tb)
}
\arguments{
\item{train_tb}{a tibble with time series data.}
}
\value{
a list of ggplot2 objects.
}
\description{
Produce figures that help diagnosing time series for trend and seasonality.
}
