% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fit_prophet_model}
\alias{fit_prophet_model}
\title{Fit a prophet model}
\usage{
fit_prophet_model(train_tb)
}
\arguments{
\item{train_tb}{a tibble with time series data.}
}
\value{
a model object (prophet).
}
\description{
Fit a model to the data using the \code{prophet} package.
}
