% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_lm.R
\name{get_lm_r2}
\alias{get_lm_r2}
\title{Get model's R squared}
\usage{
get_lm_r2(lm_obj)
}
\arguments{
\item{lm_obj}{a linear model object (stats package).}
}
\value{
a numeric.
}
\description{
Get the R squared coefficient from a regression object.
}
