% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_plot_queimadas_forecast}
\alias{get_plot_queimadas_forecast}
\title{Plot the results of the Queimadas method}
\usage{
get_plot_queimadas_forecast(x_df, y_df, forecast_df)
}
\arguments{
\item{x_df}{a data frame. The data on which \code{y} would be estimated.}

\item{y_df}{a data frame. The data used to fit a linear model on the
temporal overlaping with \code{x_df}}

\item{forecast_df}{a data frame. The results of projecting y for the time
extent of x.}
}
\value{
a plot object (ggplot2).
}
\description{
Plot the results of applying the Queimadas method.
}
